refactored for better experimentation management
