+++ gnn/configs/lightgcn_hparams.yaml
@@ -1,0 +1,31 @@
+#data
+data:
+    dataset_size : "100k"
+    trainset_ratio : 0.75
+    seed : 10708
+    
+#model
+model:
+    model_type : "" # just to make recommenders library's checks happy
+    embed_size : 64 # the embedding dimension of users and items
+    n_layers : 3 # number of layers of the model
+    neighbor_aggregator : "degree_norm" # "degree_norm", "attention"
+    info_updater : "direct" # "direct", "linear"
+    final_node_repr : "mean" # "mean", "concat", ("weighted", "attention")
+    
+#train
+train:
+    batch_size : 1024
+    decay : 0.0001 # l2 regularization for embedding parameters
+    epochs : 50 # number of epochs for training
+    learning_rate : 0.005
+    eval_epoch : 10 # if it is not -1, evaluate the model every eval_epoch; -1 means that evaluation will not be performed during training
+    top_k : 10 # number of items to recommend when calculating evaluation metrics
+
+#show info
+#metric : "recall", "ndcg", "precision", "map"
+info:
+    save_model : True # whether to save model
+    save_epoch : 50 # if save_model is set to True, save the model every save_epoch
+    metrics : ["recall", "ndcg", "precision", "map"] # metrics for evaluation
+    ckpt_dir : ./gnn/outputs/checkpoints # directory of saved models
