--- gnn/lightgcn/outputs/100k/checkpoint
@@ -1,2 +1,0 @@
-model_checkpoint_path: "epoch_50"
-all_model_checkpoint_paths: "epoch_50"
